feat(deploy): add Docker/systemd/install, CI release, and repo hygiene Ship deployment artifacts: multi-stage Dockerfile, docker-compose, systemd units, and a checksum-verifying install.sh (dry-run/uninstall), documented in deploy/README.md. Add a tag-triggered release workflow (tarball + SHA256 + protected crates.io gate), Dependabot, and MIT/Apache-2.0 license files. Repo hygiene: gitignore + remove the committed runtime state file config.toml, rename init.bash -> script/mktestdirs.sh, and move the perf script to a cargo example (examples/fs_read_perf.rs, script/README.md). Refresh the README and mark the improvement plan complete.
