--- config.toml
@@ -1,1 +1,0 @@
-works = []
