--- init.bash
@@ -1,8 +1,0 @@
-mkdir ./dictest
-mkdir -p ./lower/a
-mkdir -p ./lower/b
-mkdir -p ./lower/c
-mkdir -p ./lower/d
-mkdir -p ./upper/e
-mkdir ./workerdir
-mkdir ./true_temp
